requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==23.0.0
idna==3.11
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
drf-spectacular==0.27.2
